# Do not use context manager for connection pool borrowing