Turns out at this micro-optimization level a context manager is quite expensive. It builds an object to encapsulate the context and that adds gc pressure on top of overall slowness. Removing this provides 0.6-0.8us ~10% improvement and softens gc latency spikes in production.

Before:
multithreaded: Overall: 135509.96 RPS / 7.38 us/req
singlethreaded: Overall: 137630.69 RPS / 7.27 us/req
After:
multithreaded: Overall: 151058.43 RPS / 6.62 us/req
singlethreaded: Overall: 151288.67 RPS / 6.61 us/req


Before:
![Screenshot 2023-11-06 at 12 27 19](https://github.com/RevenueCat/meta-memcache-py/assets/1917/fe390bb2-70b2-465a-af5a-fe64872dc1b3)
After:
![Screenshot 2023-11-06 at 12 27 34](https://github.com/RevenueCat/meta-memcache-py/assets/1917/6f4d03fb-63a1-4ca2-a825-a32e6cd7257b)